# 7thProject/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Tour Management API",
    "description": "This API provides endpoints for managing tours, including listing tours by continent, retrieving popular tours, and getting recommended tours for different seasons.",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Ulukbek",
      "email": "utoraliev@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://www.google.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost:8080",
      "description": "Local ENV"
    }
  ],
  "paths": {
    "/api/v1/tours/tour/{name}/booking": {
      "post": {
        "tags": [
          "tour-controller"
        ],
        "operationId": "bookTour",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tours": {
      "get": {
        "tags": [
          "tour-controller"
        ],
        "operationId": "getTours",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tour"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tours/tour/{name}": {
      "get": {
        "tags": [
          "tour-controller"
        ],
        "operationId": "getTour",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Tour"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tours/recommended/{season}": {
      "get": {
        "tags": [
          "tour-controller"
        ],
        "operationId": "getRecommendedTours",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tour"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tours/popular": {
      "get": {
        "tags": [
          "tour-controller"
        ],
        "operationId": "getPopularTours",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tour"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tours/mostVisited": {
      "get": {
        "tags": [
          "tour-controller"
        ],
        "operationId": "getMostVisitedTours",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tour"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tours/featured": {
      "get": {
        "tags": [
          "tour-controller"
        ],
        "operationId": "getFeaturedTours",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tour"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tours/continent/{continent}": {
      "get": {
        "tags": [
          "tour-controller"
        ],
        "operationId": "getToursByContinent",
        "parameters": [
          {
            "name": "continent",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tour"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tours/comments": {
      "get": {
        "tags": [
          "tour-controller"
        ],
        "operationId": "getComments",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entry": {
      "get": {
        "tags": [
          "tour-controller"
        ],
        "operationId": "getEntry",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Entry"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookingDto": {
        "type": "object",
        "properties": {
          "phoneNum": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "seatsOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nickname": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "photoComment": {
            "$ref": "#/components/schemas/PhotoComment"
          }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "phone_num": {
            "type": "string"
          },
          "seats_order": {
            "type": "integer",
            "format": "int32"
          },
          "booked": {
            "type": "boolean"
          },
          "tour": {
            "$ref": "#/components/schemas/Tour"
          }
        }
      },
      "PhotoComment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "photo_url": {
            "type": "string"
          }
        }
      },
      "PhotoTour": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "photo_url": {
            "type": "string"
          }
        }
      },
      "Tour": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "available_seats": {
            "type": "integer",
            "format": "int32"
          },
          "season": {
            "type": "string",
            "enum": [
              "WINTER",
              "SPRING",
              "SUMMER",
              "AUTUMN"
            ]
          },
          "popular": {
            "type": "boolean"
          },
          "featured": {
            "type": "boolean"
          },
          "photoTour": {
            "$ref": "#/components/schemas/PhotoTour"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "continent": {
            "type": "string"
          }
        }
      },
      "Entry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "photoTour": {
            "$ref": "#/components/schemas/PhotoTour"
          }
        }
      }
    }
  }
}