# 7thProject/src/main/java/com/example/project/dto/BookingDto.java
package com.example.project.dto;


import lombok.*;
import lombok.experimental.FieldDefaults;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@FieldDefaults(level = AccessLevel.PRIVATE)
public class BookingDto {
    String phoneNum;
    String comment;
    int seatsOrder;
}